start moving
