+++ lib/dotcom/utils/enum.ex
@@ -1,4 +1,4 @@
-defmodule Util.EnumHelpers do
+defmodule Dotcom.Utils.Enum do
   @doc """
 
   Takes an Enumerable and returns a list with the first and last items tagged
